Update Models
Totals and discounts were previously long types. This causes math errors when calculating the discounts on the line items and sales. Switched them over to double types for better precision.

+++ LightspeedTakeHome/Models/Sale.cs
@@ -5,8 +5,8 @@
         public long Id { get; set; }
         //Making the following required because you can't have a sale without a LineItem. 
         public required List<LineItem> LineItems {get; set;}
-        public long Total {get; set;}
-        public long? SaleDiscount { get; set; }
+        public double Total {get; set;}
+        public double? SaleDiscount { get; set; }
     }
     
     public class LineItem
@@ -15,7 +15,7 @@
         public required long ProductId { get; set; } // Using this to pull the Product from the DB.
         public required long Quantity {get; set;}
         public Product? ProductForSale { get; set; } // This is the product that is being sold (fetched with ProductID), used for name and price data.
-        public long TotalCost { get; set; }
-        public long LineItemDiscount { get; set; }
+        public double TotalCost { get; set; }
+        public double LineItemDiscount { get; set; }
     }
 }
